requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
sqlalchemy
pymysql
passlib[bcrypt]
python-jose[cryptography]
email-validator
pydantic-settings
python-multipart